requirements/base.txt
# Production dependencies
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
pydantic>=2.0.0
pdfplumber>=0.10.0
mlflow>=3.1.0
openai>=1.0.0
pyspark==3.5.*
databricks-connect==15.4.*
databricks-sdk>=0.30.0
pyyaml>=6.0
ipython>=8.0.0
python-Levenshtein>=0.27.1
python-dotenv>=1.0.0 